wifi: mt76: mt7996: Use vif_to_dev when needed
There are cases where the deflink of a vif may not be initialized, and
therefore the phy in the vif_link struct may be null. This makes it
difficult to get the dev. Instead, use the vif_to_dev function to safely
obtain the dev.

CT-TODO: Squash into original introductions (have fun future me!)
Signed-off-by: Dylan Eskew <dylan.eskew@candelatech.com>